Added third party library libbacktrace
libbacktrace can be used to generate pretty stack trace. It's
part of GCC, so a shell script get_libbt.sh is needed to fetch
a copy of it and necessary supporting files from the official
GCC svn repository.
Note, although libbacktrace is released under BSD license, the
supporting files fetched by get_libbt.sh are probably released
under GPLv3. Because of these GPL'ed supporting files a major
patch is needed in order to release the whole project under
BSD license or a similar one. Those supporting files are mostly
build scripts and header files, so it's not too difficult to
get rid of them.